C implementation: Split out C++ interface and make it configurable.
Prepare for pure C implementation.

Applications should be able to include thingset.h and get what they need
independent whether they are C or C++ based. Start by spliting the interface
and making it configurable.

The default interface is C++. No functional changes.
